% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{read_assignment_file}
\alias{read_assignment_file}
\title{Read the data from the ASSIGNMENTS.csv file}
\usage{
read_assignment_file(dir = ".", example = FALSE)
}
\description{
Read the data from the ASSIGNMENTS.csv file
}

